# api/openapi.yaml
openapi: 3.1.0
info:
  title: GrantSupport Engine API
  version: 0.1.0-beta.3
  description: >
    GrantSupport is a delegated support-access authentication and cryptographic audit engine.
    It enables multi-tenant SaaS applications to issue cryptographically signed, time-bounded,
    single-use support access tokens to vendor support agents with tamper-evident audit logging.
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: https://api.grantsupport.local/v1
    description: Production HTTPS Reverse Proxy
  - url: http://127.0.0.1:8080
    description: Local Standalone Development Server
paths:
...
  /api/v1/audit/events:
    get:
      operationId: getAuditEvents
      summary: Query Cryptographic Audit Events
      description: >
        Called by a customer administrator to query paginated cryptographic audit events strictly scoped to their tenant.
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Number of audit records to return (1-100).
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Pagination offset.
        - name: actor_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter events by actor UUID.
        - name: event_type
          in: query
          required: false
          schema:
            type: string
          description: Filter events by event type string.
      responses:
        '200':
          description: Paginated list of audit events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEventsResponse'
        '400':
          $ref: '#/components/responses/RFC7807BadRequest'
        '401':
          $ref: '#/components/responses/RFC7807Unauthorized'
        '500':
          $ref: '#/components/responses/RFC7807InternalError'
  /api/v1/audit/verify:
    post:
      operationId: verifyAuditChain
      summary: Verify Cryptographic Audit Chain
      description: >
        Called by a customer administrator to cryptographically verify the unbroken SHA-256 sequential hash chain of all historical audit events for their tenant.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Audit ledger verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditVerifyResponse'
        '401':
          $ref: '#/components/responses/RFC7807Unauthorized'
        '500':
          $ref: '#/components/responses/RFC7807InternalError'
  /api/v1/access-requests:
    post:
      operationId: createAccessRequest
      summary: Submit Delegated Support Access Request
      description: >
        Called by a vendor support agent to submit an in-band Just-In-Time access request to a customer organization.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccessRequestPayload'
      responses:
        '201':
          description: Access request created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessRequest'
        '400':
          $ref: '#/components/responses/RFC7807BadRequest'
        '401':
          $ref: '#/components/responses/RFC7807Unauthorized'
        '429':
          $ref: '#/components/responses/RFC7807TooManyRequests'
        '500':
          $ref: '#/components/responses/RFC7807InternalError'
    get:
      operationId: listAccessRequests
      summary: List Access Requests
      description: >
        Queries paginated access requests for the authenticated tenant. Support agents view only their own requests.
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [PENDING, APPROVED, REJECTED, CANCELLED, EXPIRED]
          description: Filter by request status.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of access requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAccessRequestsResponse'
        '401':
          $ref: '#/components/responses/RFC7807Unauthorized'
        '500':
          $ref: '#/components/responses/RFC7807InternalError'
  /api/v1/access-requests/{id}:
    get:
      operationId: getAccessRequest
      summary: Get Access Request Details
      description: >
        Retrieves detailed information for a specific access request with strict tenant isolation.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Unique access request UUID.
      responses:
        '200':
          description: Access request details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessRequest'
        '401':
          $ref: '#/components/responses/RFC7807Unauthorized'
        '404':
          description: Request not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/RFC7807ProblemDetails'
        '500':
          $ref: '#/components/responses/RFC7807InternalError'
  /api/v1/access-requests/{id}/approve:
    post:
      operationId: approveAccessRequest
      summary: Approve Access Request & Issue One-Time Token
      description: >
        Called by a customer administrator to approve a pending access request. Atomically creates a single-use SupportGrant and returns the rawToken exactly once in the response. Self-approval is strictly forbidden.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Unique access request UUID.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveAccessRequestPayload'
      responses:
        '200':
          description: Access request approved and raw single-use token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApproveAccessRequestResponse'
        '400':
          $ref: '#/components/responses/RFC7807BadRequest'
        '401':
          $ref: '#/components/responses/RFC7807Unauthorized'
        '403':
          description: Self-approval forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/RFC7807ProblemDetails'
        '404':
          description: Request not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/RFC7807ProblemDetails'
        '409':
          description: Request already resolved
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/RFC7807ProblemDetails'
        '410':
          description: Request expired
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/RFC7807ProblemDetails'
        '429':
          $ref: '#/components/responses/RFC7807TooManyRequests'
        '500':
          $ref: '#/components/responses/RFC7807InternalError'
  /api/v1/access-requests/{id}/reject:
    post:
      operationId: rejectAccessRequest
      summary: Reject Access Request
      description: >
        Called by a customer administrator to deny a pending access request with an explanation.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectAccessRequestPayload'
      responses:
        '200':
          description: Access request rejected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardSuccessResponse'
        '400':
          $ref: '#/components/responses/RFC7807BadRequest'
        '401':
          $ref: '#/components/responses/RFC7807Unauthorized'
        '404':
          description: Request not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/RFC7807ProblemDetails'
        '409':
          description: Request already resolved
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/RFC7807ProblemDetails'
        '500':
          $ref: '#/components/responses/RFC7807InternalError'
  /api/v1/access-requests/{id}/cancel:
    post:
      operationId: cancelAccessRequest
      summary: Cancel Access Request
      description: >
        Called by the requesting agent or a customer administrator to cancel a pending access request.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Access request cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardSuccessResponse'
        '401':
          $ref: '#/components/responses/RFC7807Unauthorized'
        '403':
          description: Unauthorized cancellation
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/RFC7807ProblemDetails'
        '404':
          description: Request not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/RFC7807ProblemDetails'
        '409':
          description: Request already resolved
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/RFC7807ProblemDetails'
        '500':
          $ref: '#/components/responses/RFC7807InternalError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 signed JSON Web Token carrying tenant and role claims.
  schemas:
...
    AuditEventItem:
      type: object
      required:
        - id
        - institution_id
        - actor_id
        - event_type
        - description
        - hash_chain
        - created_at
      properties:
        id:
          type: string
          format: uuid
          example: 01a003f4-5678-7123-9abc-def012345678
        institution_id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        actor_id:
          type: string
          format: uuid
          example: 7f4c935b-16d7-4f9e-a8f2-39c4a852b719
        event_type:
          type: string
          example: SUPPORT_ACCESS_GRANTED
        description:
          type: string
          example: Support access grant created for duration 60 minutes
        hash_chain:
          type: string
          description: Sequential SHA-256 hash linking this event immutably to the preceding audit event.
          example: 9f83a8b9487c6e12e2057639f28d8442c8d2345e89a1b2c3d4e5f6a7b8c9d0e1
        created_at:
          type: string
          format: date-time
          example: 2026-08-15T12:00:00.000000Z
    AuditEventsResponse:
      type: object
      required:
        - success
        - events
        - limit
        - offset
      properties:
        success:
          type: boolean
          example: true
        events:
          type: array
          items:
            $ref: '#/components/schemas/AuditEventItem'
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0
    AuditVerifyResponse:
      type: object
      required:
        - success
        - valid
        - institution_id
        - message
      properties:
        success:
          type: boolean
          example: true
        valid:
          type: boolean
          example: true
        institution_id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        message:
          type: string
          example: Audit log cryptographic hash chain is valid and unbroken.
    StandardSuccessResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully.
    AccessRequest:
      type: object
      required:
        - id
        - institutionId
        - requesterId
        - reason
        - requestedDurationMinutes
        - requestedScope
        - status
        - expiresAt
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          example: 01a003f4-5678-7123-9abc-def012345678
        institutionId:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        requesterId:
          type: string
          format: uuid
          example: 7f4c935b-16d7-4f9e-a8f2-39c4a852b719
        targetService:
          type: string
          example: billing-service
        reason:
          type: string
          example: Investigating invoice discrepancy on customer ticket #4829
        requestedDurationMinutes:
          type: integer
          example: 60
        approvedDurationMinutes:
          type: integer
          example: 45
        requestedScope:
          type: string
          example: billing:read
        approvedScope:
          type: string
          example: billing:read
        requestedIps:
          type: array
          items:
            type: string
          example: ["198.51.100.4"]
        approvedIps:
          type: array
          items:
            type: string
          example: ["198.51.100.4"]
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED, CANCELLED, EXPIRED]
          example: PENDING
        expiresAt:
          type: string
          format: date-time
          example: 2026-08-16T12:00:00Z
        approvedById:
          type: string
          format: uuid
        approvedAt:
          type: string
          format: date-time
        rejectedById:
          type: string
          format: uuid
        rejectionReason:
          type: string
        rejectedAt:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
        supportGrantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
          example: 2026-08-15T12:00:00Z
    CreateAccessRequestPayload:
      type: object
      required:
        - reason
        - durationMinutes
      properties:
        targetService:
          type: string
          maxLength: 128
          example: billing-service
        reason:
          type: string
          minLength: 3
          maxLength: 1000
          example: Investigating invoice discrepancy on customer ticket #4829
        durationMinutes:
          type: integer
          minimum: 1
          maximum: 1440
          example: 60
        scope:
          type: string
          maxLength: 64
          example: billing:read
        whitelistedIps:
          type: array
          items:
            type: string
          example: ["198.51.100.4"]
    ApproveAccessRequestPayload:
      type: object
      properties:
        durationMinutes:
          type: integer
          minimum: 1
          maximum: 1440
          example: 45
        scope:
          type: string
          maxLength: 64
          example: billing:read
        whitelistedIps:
          type: array
          items:
            type: string
          example: ["198.51.100.4"]
    RejectAccessRequestPayload:
      type: object
      required:
        - rejectionReason
      properties:
        rejectionReason:
          type: string
          minLength: 3
          maxLength: 1000
          example: Ticket #4829 has already been resolved by support engineering.
    ApproveAccessRequestResponse:
      type: object
      required:
        - success
        - requestId
        - status
        - grantId
        - rawToken
        - expiresAt
      properties:
        success:
          type: boolean
          example: true
        requestId:
          type: string
          format: uuid
        status:
          type: string
          example: APPROVED
        grantId:
          type: string
          format: uuid
        rawToken:
          type: string
          description: Single-use high-entropy bearer token returned strictly once upon approval and never persisted unhashed.
          example: 550e8400-e29b-41d4-a716-446655440000_f92b4c10e82a4d70923f1b0a...
        expiresAt:
          type: string
          format: date-time
    ListAccessRequestsResponse:
      type: object
      required:
        - requests
        - count
      properties:
        requests:
          type: array
          items:
            $ref: '#/components/schemas/AccessRequest'
        count:
          type: integer
          example: 1
    RFC7807ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type.
          example: https://grantsupport.io/errors/validation-failed
        title:
          type: string
          description: A short, human-readable summary of the problem type.
          example: VALIDATION_FAILED
        status:
          type: integer
          description: The HTTP status code generated by the origin server.
          example: 400
        detail:
          type: string
          description: A sanitized human-readable explanation specific to this occurrence of the problem.
          example: The request payload failed schema validation constraints.
        instance:
          type: string
          description: A URI reference identifying the specific occurrence of the problem.
          example: /api/v1/auth/support/grant
        code:
          type: string
          description: Machine-readable application error code.
          example: ERR_INVALID_INPUT
        correlation_id:
          type: string
          description: Unique correlation identifier for tracing.
          example: 01a003f4-5678-7123-9abc-def012345678
        validation_errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: durationMinutes
              message:
                type: string
                example: durationMinutes must be at least 1
  responses:
    RFC7807BadRequest:
      description: Bad Request (Invalid input or constraint violation)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/RFC7807ProblemDetails'
    RFC7807Unauthorized:
      description: Unauthorized (Invalid, expired, or revoked token)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/RFC7807ProblemDetails'
    RFC7807TooManyRequests:
      description: Too Many Requests (Rate limit exceeded)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/RFC7807ProblemDetails'
    RFC7807InternalError:
      description: Internal Server Error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/RFC7807ProblemDetails'